test: fold test_schema_version_sync into test_schema_versioning
The sample-config sync guard and reference-metadata example test move
unchanged into test_schema_versioning.py, next to the older
TestSampleConfig cousin they tighten. Also refresh test/README.md,
which named six test files that no longer exist.

Part of the data_model test consolidation (Tier A: merges only, no
behaviour change).
