+++ test/README.md
@@ -5,26 +5,32 @@ This directory contains the test suite for SUEWS/SuPy, organised by functionalit
 ## Test Organisation
 
 ### Core Tests (`core/`)
-Essential core functionality tests including:
+API, CLI, and utility tests plus the numerical guardrails:
 - **test_sample_output.py** - Tolerance-based reference-output validation
 - **test_fortran_state_persistence.py** - Exercises native-state isolation across an in-process order matrix
-- **test_floating_point_stability.py** - Numerical stability and reproducibility tests
-- **test_suews_simulation.py** - High-level API interface tests
+- **test_suews_simulation.py** - High-level `SUEWSSimulation` API interface tests
 - **test_supy.py** - Comprehensive test suite (runs during wheel building)
-- **test_spurious_warnings.py** - Ensures clean imports without warnings
+- **test_load.py / test_post.py** - Forcing loading and post-processing
+- **test_util_*.py** - Utility modules (atmospheric, OHM, ERA5, surface conductance)
+- **test_cli_*.py** - CLI entry-point behaviour (run, validation, conversion)
+- CI/tooling self-tests (**test_pytest_ci_metrics.py**, **test_ci_run_metrics.py**, **test_ci_phase_metrics.py**, **test_scheduler_benchmark.py**, **test_check_schema_version_bump.py**, **test_audit_python_startup.py**) - guard the repo's instrumentation and lint gates rather than the model
 
 ### Data Model Tests (`data_model/`)
 Configuration and data model validation tests:
+- **test_validation.py** - Validator behaviour: physics-option dependencies, phase-B science checks, forcing validation, DLS/irrigation rules
+- **test_yaml_processing.py** - Three-phase YAML processor pipeline (Phase A uptodate, Phase B science check, Phase C Pydantic validation) plus precheck helpers
 - **test_data_model.py** - Data model structure and conversion tests
-- **test_precheck.py** - Pre-run validation and checks
-- **test_conditional_validation.py** - Physics option compatibility validation
-- **test_validation_topdown.py** - Top-down configuration validation
-- **test_validation_utils.py** - Validation utility functions
-- **test_flexible_refvalue_clean.py** - RefValue wrapper functionality
+- **test_physics_options.py** - Physics-option input forms: flat codes, nested family tags, orthogonal form, readable names
+- **test_renames.py** - Rename registries: Pydantic field renames and DataFrame column renames, dual-read helpers, Rust bridge alignment
+- **test_yaml_roundtrip.py** - `SUEWSConfig` YAML I/O: from_yaml error handling and drift hints, to_yaml round-trips, serialisation warnings
+- **test_yaml_upgrade.py** - Schema migration handlers (includes `TestNoSilentFieldDrops`)
+- **test_schema_versioning.py** - Schema version constant, lineage, and sample-config sync
+- **test_release_compat.py** - Vendored release fixtures round-trip through the current validator
 
 ### Physics Tests (`physics/`)
 Scientific and physics validation tests:
 - **test_core_physics.py** - Physical consistency checks (runs during wheel building)
+- Scheme-specific guardrails: anthropogenic heat, attribution, DyOHM building, irrigation wiring, Macdonald roughness, OHM coefficient blending, RSL profiles, SPARTACUS
 
 ### I/O Tests (`io_tests/`)
 Input/output and data handling tests:
@@ -32,8 +38,21 @@ Input/output and data handling tests:
 - **test_save_supy.py** - Shared output-saving backend
 - **test_resample_output.py** - Output resampling capabilities
 - **test_dailystate_output.py** - Daily state output handling
-- **test_forcing_file_list.py** - Forcing file list handling
+- **test_forcing_interpolation.py / test_named_column_forcing.py** - Forcing handling
+- **test_output_layout_parity.py** - Output layout parity across backends
 - **test_yaml_annotation.py** - YAML annotation features
+
+### CLI Command Tests (`cmd/`)
+Entry points under `suews` (init-case, validate-config, inspect-config, diagnose-run, compare-runs, summarise-output, metrics, JSON envelope contract, knowledge CLI).
+
+### MCP Server Tests (`mcp/`)
+One file per MCP tool/resource plus protocol handshake, packaging manifests, and CLI smoke tests. The one-file-per-tool layout is deliberate.
+
+### Documentation Tests (`docs/`)
+Docs-drift guards (site and docs sources), bibliography audit, and the data-model RST generator.
+
+### Knowledge Pack Tests (`knowledge/`)
+Knowledge pack integrity (`test_pack.py`).
 
 ### UMEP/QGIS Tests (`umep/`)
 UMEP plugin compatibility tests (Windows + Python 3.12 target, GH-901):
@@ -54,7 +73,8 @@ lane is enough for this repository's plugin-facing compatibility checks.
 Test data and resources:
 - **benchmark1/** - Benchmark test configuration and data
 - **data_test/** - Sample data for various tests
-- **precheck_testcase/** - Test cases for precheck functionality
+- **release_configs/** - Vendored per-release YAML shapes for migration regression
+- **legacy_tables/** - Vendored legacy table-based inputs
 
 ## Running Tests
 
@@ -181,7 +201,10 @@ When adding new tests:
 1. Place them in the appropriate category directory
 2. Follow the existing naming convention: `test_<functionality>.py`
 3. Use descriptive test names that explain what is being tested
-4. Add docstrings to explain complex test logic
-5. Update this README if adding a new test category
+4. Prefer `pytest.mark.parametrize` over copy-pasting a test body with
+   different inputs — symmetric cases (e.g. wall/roof, albedo/emissivity)
+   belong in one parametrised test
+5. Add docstrings to explain complex test logic
+6. Update this README if adding a new test category
 
 For detailed testing approach, see docstrings in test files or `docs/source/contributing/testing_guide.rst`.
